🚚 rename github project projector
