+++ marketplace-domain/src/contribution/projectors/mod.rs
@@ -4,5 +4,5 @@ pub use github_contribution::GithubContributionProjector;
 mod application;
 pub use application::ApplicationProjector;
 
-mod project;
-pub use project::ProjectProjector;
+mod github_project;
+pub use github_project::GithubProjectProjector;
