--- marketplace-domain/src/contribution/projectors/project/mod.rs
@@ -1,6 +1,0 @@
-#[allow(clippy::module_inception)]
-mod project;
-pub use project::*;
-
-#[cfg(test)]
-mod test;
