+++ marketplace-indexer/src/lib.rs
@@ -52,7 +52,7 @@ fn build_event_observer(
 
 	let contribution_projector = GithubContributionProjector::new(database.clone(), github.clone());
 	let application_projector = ApplicationProjector::new(database.clone());
-	let project_projector = ProjectProjector::new(github.clone(), database.clone());
+	let project_projector = GithubProjectProjector::new(github.clone(), database.clone());
 	let project_member_projector = ProjectMemberProjector::new(database.clone());
 	let contributor_projector = ContributorWithGithubDataProjector::new(github, database.clone());
 	let lead_contributors_projector = LeadContributorProjector::new(database.clone());
